auto_rx/requirements.txt
python-dateutil
flask
flask-socketio
numpy
requests
semver
simple-websocket
